+++ src/main/java/raisetech/StudentManagement/StudentManagementApplication.java
@@ -1,33 +1,14 @@
 package raisetech.StudentManagement;
 
-import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-import raisetech.StudentManagement.data.Student;
-import raisetech.StudentManagement.repository.StudentRepository;
 
 @SpringBootApplication
-@RestController
 public class StudentManagementApplication {
-
-	@Autowired
-	private StudentRepository repository;
 
 	public static void main(String[] args) {
 		SpringApplication.run(StudentManagementApplication.class, args);
 	}
 
-	@GetMapping("/studentList")
-	public List<Student> getStudentList() {
-		return repository.searchStudent();
-	}
-
-	@GetMapping("/studentsCourseList")
-	public List<Student> getStudentsCourseList(){
-		return repository.searchStudentCourse();
-	}
 
 }
