+++ src/main/java/raisetech/StudentManagement/controller/StudentController.java
@@ -1,0 +1,32 @@
+package raisetech.StudentManagement.controller;
+
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+import raisetech.StudentManagement.data.Student;
+import raisetech.StudentManagement.data.StudentCourse;
+import raisetech.StudentManagement.service.StudentService;
+
+@RestController
+public class StudentController {
+
+  private StudentService service;
+
+  @Autowired
+  public StudentController(StudentService service) {
+    this.service = service;
+  }
+
+  @GetMapping("/studentList")
+  public List<Student> getStudentList() {
+    return service.searchStudentList();
+  }
+
+  @GetMapping("/studentsCourseList")
+  public List<StudentCourse> getStudentsCourseList(){
+    return service.searchStudentsCourseList();
+  }
+
+
+}
