+++ src/main/java/raisetech/StudentManagement/repository/StudentRepository.java
@@ -4,7 +4,13 @@ import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
 import raisetech.StudentManagement.data.Student;
+import raisetech.StudentManagement.data.StudentCourse;
 
+/**
+ * 受講生情報を扱うリポジトリ
+ *
+ * 全件検索や単一条件での検索、コース情報の検索が行えるクラスです
+ */
 @Mapper
 public interface StudentRepository {
 
@@ -12,6 +18,6 @@ public interface StudentRepository {
   List<Student> searchStudent();
 
   @Select("SELECT * FROM students_courses")
-  List<Student> searchStudentCourse();
+  List<StudentCourse> searchStudentCourse();
 
 }
