+++ src/main/java/raisetech/StudentManagement/service/StudentService.java
@@ -1,0 +1,27 @@
+package raisetech.StudentManagement.service;
+
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import raisetech.StudentManagement.data.Student;
+import raisetech.StudentManagement.data.StudentCourse;
+import raisetech.StudentManagement.repository.StudentRepository;
+
+@Service
+public class StudentService {
+
+  private StudentRepository repository;
+
+  @Autowired
+  public StudentService(StudentRepository repository) {
+    this.repository = repository;
+  }
+
+  public List<Student> searchStudentList() {
+    return repository.searchStudent();
+  }
+
+  public List<StudentCourse> searchStudentsCourseList(){
+    return repository.searchStudentCourse();
+  }
+}
